Revert "move style/ and lib/ to src/"
