+++ lib/VRLoader.js
