+++ src/components/camera.js
@@ -1,5 +1,5 @@
 var registerComponent = require('../core/component').registerComponent;
-var THREE = require('../lib/three');
+var THREE = require('../../lib/three');
 
 module.exports.Component = registerComponent('camera', {
   schema: {
