+++ src/components/sound.js
@@ -1,7 +1,7 @@
 var debug = require('../utils/debug');
 var diff = require('../utils').diff;
 var registerComponent = require('../core/component').registerComponent;
-var THREE = require('../lib/three');
+var THREE = require('../../lib/three');
 
 var warn = debug('components:sound:warn');
