+++ src/components/wasd-controls.js
@@ -1,5 +1,5 @@
 var registerComponent = require('../core/component').registerComponent;
-var THREE = require('../lib/three');
+var THREE = require('../../lib/three');
 
 var MAX_DELTA = 0.2;
