+++ src/index.js
@@ -4,11 +4,11 @@ require('present');  // Polyfill `performance.now()`.
 // TODO: Extract to aframe-primitives.
 // HTML Imports polyfill must come before everything else.
 if (!('import' in document.createElement('link'))) {
-  require('./lib/vendor/HTMLImports');
+  require('../lib/vendor/HTMLImports');
 }
 
-require('./style/aframe-core.css');
-require('./style/rStats.css');
+require('../style/aframe-core.css');
+require('../style/rStats.css');
 
 // Required before `AEntity` so that all components are registered.
 var AScene = require('./core/a-scene');
@@ -17,7 +17,7 @@ var debug = require('./utils/debug');
 var registerComponent = require('./core/component').registerComponent;
 var registerElement = require('./core/a-register-element');
 // Exports THREE to window so three.js can be used without alteration.
-var THREE = window.THREE = require('./lib/three');
+var THREE = window.THREE = require('../lib/three');
 var TWEEN = window.TWEEN = require('tween.js');
 
 var pkg = require('../package');
@@ -54,7 +54,6 @@ module.exports = {
   registerComponent: registerComponent,
   registerElement: registerElement,
   THREE: THREE,
-  TWEEN: TWEEN,
   utils: utils,
   version: pkg.version,
